# Faster sort