--- a/index.js
+++ b/index.js
@@ -7,6 +7,7 @@ const pullAsync = require('pull-async')
 const TypedFastBitSet = require('typedfastbitset')
 const bsb = require('binary-search-bounds')
 const multicb = require('multicb')
+const FastPriorityQueue = require('fastpriorityqueue')
 const debug = require('debug')('jitdb')
 const debugQuery = debug.extend('query')
 const Status = require('./status')
@@ -1000,22 +1001,30 @@ module.exports = function (log, indexesPath) {
     })
   }
 
+  function compareAscending(a, b) {
+    return b.timestamp > a.timestamp
+  }
+
+  function compareDescending(a, b) {
+    return a.timestamp > b.timestamp
+  }
+
   function sortedByTimestamp(bitset, descending) {
     updateCacheWithLog()
     const order = descending ? 'descending' : 'ascending'
-    if (sortedCache[order].has(bitset)) return sortedCache[order].get(bitset)
-    const timestamped = bitset.array().map((seq) => {
-      return {
+    if (sortedCache[order].has(bitset))
+      return sortedCache[order].get(bitset).clone()
+    const fpq = new FastPriorityQueue(
+      descending ? compareDescending : compareAscending
+    )
+    bitset.array().forEach((seq) => {
+      fpq.add({
         seq,
         timestamp: indexes['timestamp'].tarr[seq],
-      }
+      })
     })
-    const sorted = timestamped.sort((a, b) => {
-      if (descending) return b.timestamp - a.timestamp
-      else return a.timestamp - b.timestamp
-    })
-    sortedCache[order].set(bitset, sorted)
-    return sorted
+    sortedCache[order].set(bitset, fpq.clone())
+    return fpq
   }
 
   function getMessagesFromBitsetSlice(
@@ -1029,12 +1038,15 @@ module.exports = function (log, indexesPath) {
     seq = seq || 0
 
     const sorted = sortedByTimestamp(bitset, descending)
-    const sliced =
-      limit != null
-        ? sorted.slice(seq, seq + limit)
-        : seq > 0
-        ? sorted.slice(seq)
-        : sorted
+    const resultSize = sorted.size
+
+    const sliced = []
+    for (var i = 0; i < seq; ++i) sorted.poll()
+
+    while (!sorted.isEmpty()) {
+      sliced.push(sorted.poll())
+      if (limit !== null && sliced.length == limit) break
+    }
 
     push(
       push.values(sliced),
@@ -1046,7 +1058,7 @@ module.exports = function (log, indexesPath) {
       push.collect((err, results) => {
         cb(err, {
           results: results,
-          total: sorted.length,
+          total: resultSize,
         })
       })
     )
@@ -1054,7 +1066,7 @@ module.exports = function (log, indexesPath) {
 
   function countBitsetSlice(bitset, seq, descending) {
     if (!seq) return bitset.size()
-    else return sortedByTimestamp(bitset, descending).slice(seq).length
+    else return
   }
 
   function paginate(operation, seq, limit, descending, onlyOffset, cb) {