@arj03 2021-02-19
Need to link [this](https://github.com/ssb-ngi-pointer/ssb-db2/issues/108) db2 benchmark issue.

I'll look at the tests, was working fine locally.