@arj03 2021-02-19
Failing test:

```
[
  {
    previous: null,
    sequence: 1,
    author: '@KNsYec80tCESsq7Q/x21ujtqts78gZ/LUflLw8dqByg=.ed25519',
    timestamp: 1613719231934,
    hash: 'sha256',
    content: { type: 'post', text: '1st' },
    signature: '/OgsgJW+/DPtj/JGlOWGAkRcV4c8ibtBwfRmdlPiRZyYz1yMofLYVc5Cn346/gxuFD2pOuSo8UYZ3i8dVDBdCA==.sig.ed25519'
  },
  {
    previous: '%vONc1qtCUZp73PgM4yMrpmEFknhq+WPRjRUoI35StxA=.sha256',
    sequence: 3,
    author: '@KNsYec80tCESsq7Q/x21ujtqts78gZ/LUflLw8dqByg=.ed25519',
    timestamp: 1613719229937,
    hash: 'sha256',
    content: { type: 'post', text: '3rd' },
    signature: 'DF5p0+NheKkPQe4W8gF8im6/znglACrebabUqU9lyEdcwvjtCFwbp5nMd1TQKoHV5yHz5hvWdiLCbeRU0X2yDw==.sig.ed25519'
  },
  {
    previous: '%Z2bIU8EzqRp787LIAsoZ401RBdCachb0DR0xEQoasBY=.sha256',
    sequence: 2,
    author: '@KNsYec80tCESsq7Q/x21ujtqts78gZ/LUflLw8dqByg=.ed25519',
    timestamp: 1613719230936,
    hash: 'sha256',
    content: { type: 'post', text: '2nd' },
    signature: 'vc30K0Pd6dVSn7vDgQQla7u9TkmWNrDquka22+ltbeaolUVQyron+aZSxP9ujBHyy4eFCRZTsUFWHywIJNVsAA==.sig.ed25519'
  }
]
```

Lets see if it is better now.