@arj03 2021-02-19
Right, the problem is that we take the min of arrival and declared and if you are unlucky the arrival on 2 of them will be the same. So the test doesn't really work.