@staltz 2021-02-19
Very neat idea with a priority queue! I'll review carefully later. I might be able to help with `sortedCache` since I introduced that.

By the way, I also thought that the sorting could be done gradually. Like if we are doing paginate with pageSize=1 every time, then basically we are just looking for maximums. I know that insertion sort allows us to do it gradually, but I wonder if there are other faster algorithms that do it gradually. Then we can spread that sorting workload across the many paginate calls.