@arj03 2021-02-19
I put this into the browser and have been clicking around. Maybe it's just psychological, but this feels snappier than before. The numbers are also good. Public is now 40ms.