@arj03 2021-02-20
Should have mentioned in the other one, but with the latest changes from @staltz this sometimes goes as low as 164 (first page) and 180 (total, 2x25 messages). It is a bit all over the place, but somewhere between 164 and 240 firs the first one.