@arj03 2021-02-20
I see the same here locally. For reference Paginate 1 big index was 269ms [before](https://github.com/ssb-ngi-pointer/jitdb/pull/122) before.