@staltz 2021-02-20
I think these numbers are real, and the "paginate 1 huge index" benchmark is going through 4000 msgs, for each of those it takes roughly 10ms, sometimes less, which individually sounds like a little, but in aggregate it's indeed 20s+.