@staltz 2021-02-20
On master branch, I put some markers to measure performance of doing the cache lookup and then a simple `arr.slice()`, and individually it takes about 0.007ms.