@staltz 2021-02-20
> Right 5 * 4000. So the question is, is that a good benchmark to have? I mean is it a realistic case. I'm thinking something like 25 * 10 is maybe a better benchmark to have.

Let me take a look if any one of the queries I do in Manyverse looks like this. But I'm inclined to say that yes it matters, because to build ssb-threads we want to go through several `post`s as quickly as possible. All things considered, on master branch we had `1300ms` for the first public query, then `1100ms` to go through 4000 msgs, so `2400ms` in total. In this PR we have the total number adding up to `24000ms`, 10x larger.

I do like the FastPriorityQueue idea, I just think we need to be careful with it. So let's not merge just yet.