@arj03 2021-02-20
Ok, maybe you can increase the chunksize a bit?