@staltz 2021-02-20
Hmm, that's a decent idea, I wonder if I could make a pull-stream operator that makes it transparent so that I can pull one item at a time, but underneath it's actually using pagination with a large page size.

I did a quick study, see below. Sensible page sizes are in the hundreds, 500 was a sweet spot.

Total items = 20k in all of the scenarios below

| Page size | Number of pages | Duration |
|-----|------|----------|
| 1 | 20000 | 36532ms |
| 5 | 4000 | 18500ms |
| 10 | 2000 | 8870ms |
| 20 | 1000 | 4557ms |
| 50 | 400 | 2206ms |
| 100 | 200 | 1171ms |
| 200 | 100 | 657ms |
| 500 | 40 | 398ms |
| 1000 | 20 | 349ms |
| 2000 | 10 | 320ms |
| 5000 | 4 | 291ms |
| 20000 | 1 | 246ms |