@staltz 2021-02-20
I approve this PR, but I'd like to add a benchmark that uses pageSize=5 and another that uses pageSize=1000, just so that we keep tracking the performance of these different ranges of values.