@arj03 2021-02-20
Please do, not in a super rush to get this merged.

It's a bit of a trade-of here. I like the idea of a chunked reader, I wonder if we could do something that would be general enough to work in the async iterator case as well. Almost sounds like someone must have thought of that problem before :thinking: