@staltz 2021-02-20
> I like the idea of a chunked reader, I wonder if we could do something that would be general enough to work in the async iterator case as well. Almost sounds like someone must have thought of that problem before thinking

Now that I think about it, I don't think we need a new operator, we can just do

```js
pull(
  query(
    fromDB(db),
    and(equal(seekType, 'post', { indexType: 'type' })),
    paginate(PAGESIZE),
    toPullStream()
  ),
  pull.take(NUMPAGES),
  pull.map(page => pull.values(page)), // this
  pull.flatten(),                      // and this
  pull.drain(op)
) 
```

For async iterators, people can do it manually or use an async iterable library such as https://github.com/ReactiveX/IxJS, we don't need to cover those use cases, because JITDB's responsibility ends after the `toPullStream()` or `toAsyncIterator()`.